obstacle made 2.5 expand, time print statements

+++ hybrid_a_star_ws/src/Hybrid_A_Star/include/hybrid_a_star/hybrid_a_star_flow.h
@@ -86,7 +86,8 @@ private:
     std::deque<geometry_msgs::PoseWithCovarianceStampedPtr> init_pose_deque_;
     std::deque<geometry_msgs::PoseStampedPtr> goal_pose_deque_;
     std::deque<nav_msgs::OccupancyGridPtr> costmap_deque_;
-    std::vector<int> vals_accessed;
+    // std::vector<int> vals_accessed;
+    std::vector<double> vals_accessed;
 
     std::vector<double> x_values;
     std::vector<double> y_values;
